Play It Safe: Manage Security Risks
This is my second certificate studying the Google Cybersecurity Professional. I learned a lot in this course and a lot that was taught made me understand information technology in a whole different way. Security is not only for the on-site or on-prem services but cloud computing too. I got to know about tools that is used for data analysis on events and logs happening in the company. Tools like Splunk.
1 . I learned to Identify the primary threats, risks, and vulnerabilities to business operations.
2. Examined how organizations use security frameworks and controls to protect business operations.
3. Defined commonly used Security Information and Event Management (SIEM) tools like Splunk.
4. Use a playbook to respond to threats, risks, and vulnerabilities.